Changed behavior of the GUI.
3DS/Switch: Changed the behavior of the GUI with respect to buttons A and Y.

Pressing the A button selects the backup list, and pressing A again with "New..." or a backup selected will create or restore a backup.

Pressing Y will select the current title but deactivate/reset the backup list. It will also "enable" the backup touch button. Creating a backup will then disable the backup touch button. (This makes it more clear to the user what buttons / UI elements are relevant to multi-backup.)