Changed behavior of GUI elements and buttons A and Y. (#151). Fix #150
* Changed A key behavior

Split the handling of touching the backup list and pressing the A button into two separate decisions. Also changed the A button to select items in the backup list. Selecting "New..." (index 0) performs a backup. Selecting a backup restores the backup.

* Tabs to Spaces

Converted leading tabs to spaces (as per the original trunk).

* Changed behavior of the GUI.

3DS/Switch: Changed the behavior of the GUI with respect to buttons A and Y.

Pressing the A button selects the backup list, and pressing A again with "New..." or a backup selected will create or restore a backup.

Pressing Y will select the current title but deactivate/reset the backup list. It will also "enable" the backup touch button. Creating a backup will then disable the backup touch button. (This makes it more clear to the user what buttons / UI elements are relevant to multi-backup.)