Upgrade code to compile on libnx 4.2.0 (with functionality issues).
NOTE: This commit will compile, but many of the interactions do not work
properly. In particular, anything that has to do with the backup
scrollable list -- tapping on it with the touch screen, pressing 'A' to
enter list, or using 'Y' to do multiselect. I wouldn't normally submit a
PR that has malfunctioning behavior, but the amount of changes needed
just to compile was not insignificant.

The HID code was rewritten, so I did my best to port all of it forward
to the new APIs. It compiles and some interactions work, but others
will need to be fixed.

I also fixed a few other issues, such as enum name changes and such.